#961: Fix saving relations

+++ src/main/java/org/benetech/servicenet/service/impl/ServiceServiceImpl.java
@@ -1,8 +1,13 @@
 package org.benetech.servicenet.service.impl;
 
 import org.benetech.servicenet.domain.Service;
+import org.benetech.servicenet.domain.ServiceAtLocation;
+import org.benetech.servicenet.domain.ServiceTaxonomy;
 import org.benetech.servicenet.repository.ServiceRepository;
+import org.benetech.servicenet.service.EligibilityService;
+import org.benetech.servicenet.service.ServiceAtLocationService;
 import org.benetech.servicenet.service.ServiceService;
+import org.benetech.servicenet.service.ServiceTaxonomyService;
 import org.benetech.servicenet.service.dto.ServiceDTO;
 import org.benetech.servicenet.service.mapper.ServiceMapper;
 import org.slf4j.Logger;
@@ -31,9 +36,20 @@ public class ServiceServiceImpl implements ServiceService {
 
     private final ServiceMapper serviceMapper;
 
-    public ServiceServiceImpl(ServiceRepository serviceRepository, ServiceMapper serviceMapper) {
+    private final ServiceAtLocationService serviceAtLocationService;
+
+    private final ServiceTaxonomyService serviceTaxonomyService;
+
+    private final EligibilityService eligibilityService;
+
+    public ServiceServiceImpl(ServiceRepository serviceRepository, ServiceMapper serviceMapper,
+        ServiceAtLocationService serviceAtLocationService, ServiceTaxonomyService serviceTaxonomyService,
+        EligibilityService eligibilityService) {
         this.serviceRepository = serviceRepository;
         this.serviceMapper = serviceMapper;
+        this.serviceAtLocationService = serviceAtLocationService;
+        this.serviceTaxonomyService = serviceTaxonomyService;
+        this.eligibilityService = eligibilityService;
     }
 
     /**
@@ -177,6 +193,18 @@ public class ServiceServiceImpl implements ServiceService {
     @Override
     public void delete(UUID id) {
         log.debug("Request to delete Service : {}", id);
-        serviceRepository.deleteById(id);
+        Service service = findById(id).orElse(null);
+        if (service != null) {
+            for (ServiceTaxonomy st : service.getTaxonomies()) {
+                serviceTaxonomyService.delete(st.getId());
+            }
+            for (ServiceAtLocation sat : service.getLocations()) {
+                serviceAtLocationService.delete(sat.getId());
+            }
+            if (service.getEligibility() != null) {
+                eligibilityService.delete(service.getEligibility().getId());
+            }
+            serviceRepository.deleteById(id);
+        }
     }
 }
